# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FamilyBranch API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterBody {
  email: string;
  /** @minLength 8 */
  password: string;
  firstName: string;
  lastName: string;
  /** Name of the family unit to create on registration */
  unitName: string;
  /** The admin's own relationship label (e.g. Mom) */
  relationshipLabel: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface Person {
  id: string;
  firstName: string;
  lastName: string;
  photoUrl?: string | null;
  phone?: string | null;
  email?: string | null;
  addressLine1?: string | null;
  addressCity?: string | null;
  addressState?: string | null;
  addressZip?: string | null;
  addressCountry?: string | null;
  birthday?: string | null;
  showBirthYear: boolean;
  instagram?: string | null;
  facebook?: string | null;
  tiktok?: string | null;
  linkedin?: string | null;
  otherSocial?: string | null;
  relationshipLabel: string;
  familyUnitId: string;
  isAdmin: boolean;
  claimed: boolean;
  claimedAt?: string | null;
  inviteExpiresAt?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type FamilyUnitParentLinkStatus =
  (typeof FamilyUnitParentLinkStatus)[keyof typeof FamilyUnitParentLinkStatus];

export const FamilyUnitParentLinkStatus = {
  none: "none",
  pending: "pending",
  accepted: "accepted",
} as const;

export interface FamilyUnit {
  id: string;
  unitName: string;
  unitCode: string;
  parentUnitId?: string | null;
  parentLinkStatus: FamilyUnitParentLinkStatus;
  parentLinkedAt?: string | null;
  createdAt: string;
  memberCount: number;
  claimedCount: number;
}

export type PersonWithUnit = Person & {
  familyUnit: FamilyUnit;
};

export interface AuthResponse {
  person: PersonWithUnit;
  /** JWT session token */
  token: string;
}

export interface UpdatePersonBody {
  firstName?: string;
  lastName?: string;
  photoUrl?: string | null;
  phone?: string | null;
  email?: string | null;
  addressLine1?: string | null;
  addressCity?: string | null;
  addressState?: string | null;
  addressZip?: string | null;
  addressCountry?: string | null;
  birthday?: string | null;
  showBirthYear?: boolean;
  instagram?: string | null;
  facebook?: string | null;
  tiktok?: string | null;
  linkedin?: string | null;
  otherSocial?: string | null;
  relationshipLabel?: string;
}

export interface FamilyUnitSummary {
  id: string;
  unitName: string;
  unitCode: string;
  memberCount: number;
}

export type FamilyUnitWithMembers = FamilyUnit & {
  members: Person[];
};

export interface CreateFamilyUnitBody {
  unitName: string;
}

export interface UpdateFamilyUnitBody {
  unitName?: string;
}

export interface AddMemberBody {
  firstName: string;
  lastName: string;
  relationshipLabel: string;
}

export interface InviteResponse {
  inviteToken: string;
  inviteUrl: string;
  inviteExpiresAt: string;
}

export interface InviteDetails {
  personId: string;
  firstName: string;
  lastName: string;
  relationshipLabel: string;
  unitName: string;
  inviteExpiresAt: string;
}

export interface ClaimProfileBody {
  email: string;
  /** @minLength 8 */
  password: string;
}

export interface CreateLinkRequestBody {
  targetUnitId: string;
  connectorPersonId: string;
}

export type LinkRequestStatus =
  (typeof LinkRequestStatus)[keyof typeof LinkRequestStatus];

export const LinkRequestStatus = {
  pending: "pending",
  accepted: "accepted",
  declined: "declined",
} as const;

export interface LinkRequest {
  id: string;
  requestingUnitId: string;
  requestingUnitName: string;
  targetUnitId: string;
  targetUnitName: string;
  connectorPersonId: string;
  connectorPersonName: string;
  status: LinkRequestStatus;
  createdAt: string;
  respondedAt?: string | null;
}

export interface LinkRequestsResponse {
  incoming: LinkRequest[];
  outgoing: LinkRequest[];
}

export interface FamilyTreeNode {
  unitId: string;
  unitName: string;
  parentUnitId?: string | null;
  members: Person[];
  children: FamilyTreeNode[];
}

export interface FamilyTree {
  rootUnit: FamilyTreeNode;
  totalUnits: number;
  totalMembers: number;
}

export interface UnitSummary {
  unitId: string;
  unitName: string;
  totalMembers: number;
  claimedMembers: number;
  unclaimedMembers: number;
  /** Percentage of members who have claimed their profile (0–100) */
  claimRate: number;
  linkedUnits: number;
  pendingInvites: number;
  pendingLinkRequests: number;
}

export interface BirthdayEntry {
  personId: string;
  firstName: string;
  lastName: string;
  relationshipLabel: string;
  birthday: string;
  unitName: string;
  /** Days until their next birthday */
  daysUntil: number;
}

export type SearchFamilyUnitsParams = {
  q: string;
};