# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FamilyBranch API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get the currently authenticated person
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PersonWithUnit"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Persons ────────────────────────────────────────────────────
...
  /family-units:
    post:
      operationId: createFamilyUnit
      tags: [family-units]
      summary: Create a new family unit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFamilyUnitBody"
      responses:
        "201":
          description: Family unit created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FamilyUnit"
  /family-units/search:
    get:
      operationId: searchFamilyUnits
      tags: [family-units]
      summary: Search family units by name or unit code
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Matching units
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FamilyUnitSummary"
  /family-units/{unitId}:
    get:
      operationId: getFamilyUnit
      tags: [family-units]
      summary: Get a family unit with its members
      parameters:
        - $ref: "#/components/parameters/unitId"
      responses:
        "200":
          description: Family unit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FamilyUnitWithMembers"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateFamilyUnit
      tags: [family-units]
      summary: Update family unit name or details (admin only)
      parameters:
        - $ref: "#/components/parameters/unitId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFamilyUnitBody"
      responses:
        "200":
          description: Updated unit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FamilyUnit"

  # ─── Members ─────────────────────────────────────────────────────
...
  /invites/{token}:
    get:
      operationId: getInvite
      tags: [invites]
      summary: Look up an invite by token (public — no auth required)
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Invite details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InviteDetails"
        "404":
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /invites/{token}/claim:
    post:
      operationId: claimProfile
      tags: [invites]
      summary: Claim a profile via invite token
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClaimProfileBody"
      responses:
        "200":
          description: Profile claimed and logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Link Requests ────────────────────────────────────────────────
...
components:
  parameters:
...
    unitId:
      name: unitId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
...
    RegisterBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        firstName:
          type: string
        lastName:
          type: string
        unitName:
          type: string
          description: Name of the family unit to create on registration
        relationshipLabel:
          type: string
          description: The admin's own relationship label (e.g. Mom)
      required: [email, password, firstName, lastName, unitName, relationshipLabel]
    LoginBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        person:
          $ref: "#/components/schemas/PersonWithUnit"
        token:
          type: string
          description: JWT session token
      required: [person, token]

    # ─── Person ────────────────────────────────────────────────────
    Person:
      type: object
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        photoUrl:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        addressLine1:
          type: string
          nullable: true
        addressCity:
          type: string
          nullable: true
        addressState:
          type: string
          nullable: true
        addressZip:
          type: string
          nullable: true
        addressCountry:
          type: string
          nullable: true
        birthday:
          type: string
          nullable: true
        showBirthYear:
          type: boolean
        instagram:
          type: string
          nullable: true
        facebook:
          type: string
          nullable: true
        tiktok:
          type: string
          nullable: true
        linkedin:
          type: string
          nullable: true
        otherSocial:
          type: string
          nullable: true
        relationshipLabel:
          type: string
        familyUnitId:
          type: string
          format: uuid
        isAdmin:
          type: boolean
        claimed:
          type: boolean
        claimedAt:
          type: string
          format: date-time
          nullable: true
        inviteExpiresAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - firstName
        - lastName
        - relationshipLabel
        - familyUnitId
        - isAdmin
        - claimed
        - showBirthYear
        - createdAt
        - updatedAt
    PersonWithUnit:
      allOf:
        - $ref: "#/components/schemas/Person"
        - type: object
          properties:
            familyUnit:
              $ref: "#/components/schemas/FamilyUnit"
          required: [familyUnit]
...
    FamilyUnit:
      type: object
      properties:
        id:
          type: string
          format: uuid
        unitName:
          type: string
        unitCode:
          type: string
        parentUnitId:
          type: string
          format: uuid
          nullable: true
        parentLinkStatus:
          type: string
          enum: [none, pending, accepted]
        parentLinkedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        memberCount:
          type: integer
        claimedCount:
          type: integer
      required:
        - id
        - unitName
        - unitCode
        - parentLinkStatus
        - createdAt
        - memberCount
        - claimedCount
    FamilyUnitSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        unitName:
          type: string
        unitCode:
          type: string
        memberCount:
          type: integer
      required: [id, unitName, unitCode, memberCount]
    FamilyUnitWithMembers:
      allOf:
        - $ref: "#/components/schemas/FamilyUnit"
        - type: object
          properties:
            members:
              type: array
              items:
                $ref: "#/components/schemas/Person"
          required: [members]
    CreateFamilyUnitBody:
      type: object
      properties:
        unitName:
          type: string
      required: [unitName]
    UpdateFamilyUnitBody:
      type: object
      properties:
        unitName:
          type: string

    # ─── Members ───────────────────────────────────────────────────
...
    InviteDetails:
      type: object
      properties:
        personId:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        relationshipLabel:
          type: string
        unitName:
          type: string
        inviteExpiresAt:
          type: string
          format: date-time
      required: [personId, firstName, lastName, relationshipLabel, unitName, inviteExpiresAt]
    ClaimProfileBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
      required: [email, password]

    # ─── Link Requests ─────────────────────────────────────────────
...